requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
cx-Oracle==6.3.1
Django==2.0.13
django-cors-headers==2.5.3
django-extensions==2.1.6
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.8
joblib==0.13.2
json-tricks==3.13.1
ldap3==2.6
nltk==3.4.4
numpy==1.16.4
pandas==0.22.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pycryptodome==3.8.1
PyJWT==1.7.1
pyOpenSSL==19.0.0
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1
requests==2.21.0
scikit-learn==0.20.4
scipy==1.2.2
six==1.12.0
sklearn==0.0
typing==3.6.6
urllib3==1.24.2
